change miles type to float

+++ RunningFeverApp/Models/Run.cs
@@ -5,7 +5,7 @@ public class Run
     [Key]
     public int Id { get; set; } 
     public DateTime StartTime {get; set;}
-    public int Miles  {get; set;}
+    public float Miles  {get; set;}
     public string? Route {get; set;}
     public string? Note {get; set;}
